Began Work on MemoryManagement
Types.d.ts - Added AllyConstant
Made some RoomMemory properties optional
