+++ src/Helpers/RoomHelper.ts
@@ -1,3 +1,6 @@
+import { MemoryApi } from "Api/Memory.Api";
+import { MemoryHelper } from "./MemoryHelper";
+
 // helper functions for rooms
 export class RoomHelper {
 
@@ -26,6 +29,7 @@ export class RoomHelper {
      * @param room the room we want to check
      */
     public static isSourceKeeperRoom(room: any): boolean {
+        //Just do that regex thing to check, we have it in our old code
         return false;
     }
 
@@ -44,129 +48,7 @@ export class RoomHelper {
     public static inTravelRange(room: Room): boolean {
         return false;
     }
-
-    /**
-     * return all the objects of a specified type in the room
-     * @param room the room we want to get objects from
-     * @param objectConst [optional] the structure we want
-     */
-    public static getObjectsInRoom(room: Room, objectConst?: StructureConstant): StringMap {
-        
-        let allObjectsID: StringMap;
-
-        // if no structure was specified, get all the structures
-        if(!objectConst){
-            allObjectsID = room.memory.structures;
-        }
-        else{ // if structure specified, get by that
-            allObjectsID = room.memory.structures[objectConst];
-        }
-
-        const allObjects: StringMap = _.map(allObjectsID, (o) => Game.getObjectById(o));
-
-        if(allObjects !== undefined){
-            return allObjects;
-        }
-
-        // if not, throw memory not set exception
-        throw new Error(`Memory not set for structure ${objectConst} in room ${room.name}.`);
-    }
-
-    /**
-     * return all the objects of a specified type in the room by a filter function
-     * @param room the room we want to get objects from
-     * @param filterFunction the filter for the objects
-     * @param objectConst [optional] the structure we want
-     */
-    public static getObjectsInRoomBy(room: Room, filterFunction: (o: any) => boolean, objectConst?: StructureConstant): StringMap {
-
-        let allObjects: StringMap;
-
-        // if no structure was specified, get all of the structures
-        if(!objectConst){
-            allObjects = this.getObjectsInRoom(room);
-        }
-        else{ // if structure was specified, get by that
-            allObjects = this.getObjectsInRoom(room, objectConst);
-        }
-
-        return _.filter(allObjects, filterFunction);
-    }
-
-    // note : it wouldn't let be put RoleConstant for the creepConst type on some of these, so i chagned all to make it consistant
-    /**
-     * get all the creeps in the room 
-     * @param room the room we want to get creeps from
-     * @param creepConst [optional] the RoleConstant 
-     */
-    public static getCreepsInRoom(room: Room, creepConst?: any): StringMap {
-
-        let allCreeps: StringMap;
-
-        // if no role was specified, get all of the creeps
-        if(!creepConst){
-            allCreeps = room.memory.creeps;
-        }
-        else{ // if role was specified, get just those roles
-            allCreeps = room.memory.creeps[creepConst];
-        }
-        
-        if(allCreeps !== undefined){
-            return allCreeps;
-        }
-
-        // if not, throw memory not set exception
-        throw new Error(`Memory not set for creep ${creepConst} in room ${room.name}.`);
-    }
-
-
-    /**
-     * get creeps in room by a filter function
-     * @param room the room we want to get creeps from
-     * @param filterFunction the filter for the creeps
-     * @param creepConst [optional] the type of creep you want (RoleConstant)
-     */
-    public static getCreepsInRoomBy(room: Room, filterFunction: (c: any) => boolean, creepConst?: any): StringMap{
-
-        let allCreeps: StringMap;
-
-        // if no role was specified, get all of the creeps
-        if(!creepConst){
-            allCreeps = this.getCreepsInRoom(room);
-        }
-        else{ // if role specified, get by that
-            allCreeps = this.getCreepsInRoom(room, creepConst);
-        }
-
-        return _.filter(allCreeps, filterFunction);
-    }
-
-    /**
-     * get number of creeps in room (can pass a filter function)
-     * @param room room we want the creeps from
-     * @param creepConst [optional] creep role we want
-     * @param filterFunction [optional] the function we want to filter by
-     */ 
-    public static getNumCreepsInRoomBy(room: Room, creepConst?: any, filterFunction?: (c: any) => boolean): number {
-
-        let allCreeps: StringMap;
-
-        // if no role was specified, get all of the creeps
-        if(!creepConst){
-            allCreeps = this.getCreepsInRoom(room);
-        }
-        else{ // if role specified, get by that
-            allCreeps = this.getCreepsInRoom(room, creepConst);
-        }
-      
-        // if no filter function provdied
-        if(!filterFunction){
-            return allCreeps.length;
-        }
-        // if filter function is provided, apply it
-        return _.filter(allCreeps, filterFunction).length;
-    }
-
+    
     /**
      * check if the object exists within a room
      * @param room the room we want to check
@@ -175,7 +57,7 @@ export class RoomHelper {
     public static isExistInRoom(room: any, objectConst: StructureConstant): boolean{
 
         // return true if any of this object exists
-        if(this.getObjectsInRoom(room, objectConst).length > 0){
+        if(MemoryHelper.getObjectsInRoom(room, objectConst).length > 0){
             return true;
         }
 
@@ -233,7 +115,7 @@ export class RoomHelper {
     }
 
     /**
-     * only returns true every ${paramter} number of ticks
+     * only returns true every ${parameter} number of ticks
      * @param ticks the number of ticks you want between executions
      */
     public static excecuteEveryTicks(ticks: number): boolean{
